+++ packages/golden_toolkit/lib/src/testing_tools.dart
@@ -192,7 +192,6 @@ Future<void> compareWithGolden(
     fail(
         'Golden tests MUST be run within a testGoldens method, not just a testWidgets method. This is so we can be confident that running "flutter test --name=GOLDEN" will run all golden tests.');
   }
-
   final shouldSkipGoldenGeneration = skip ?? GoldenToolkit.configuration.skipGoldenAssertion();
 
   final pumpAfterPrime = customPump ?? _onlyPumpAndSettle;
@@ -200,14 +199,13 @@ Future<void> compareWithGolden(
   RepaintBoundary, it should not matter */
   final actualFinder = finder ?? find.byWidgetPredicate((w) => true).first;
   final fileName = fileNameFactory(name, device);
+  final originalWindowSize = tester.binding.window.physicalSize;
 
   // This is a minor optimization and works around an issue with the current hacky implementation of invoking the golden assertion method.
   if (!shouldSkipGoldenGeneration) {
     await _primeImages(fileName, actualFinder);
   }
   await pumpAfterPrime(tester);
-
-  final originalWindowSize = tester.binding.window.physicalSize;
 
   if (autoHeight == true) {
     // Find the first scrollable element which can be scrolled vertical.