+++ packages/golden_toolkit/test/image_loading/default/default_asset_loading_test.dart
@@ -60,12 +60,19 @@ void main() {
       testWidgets('should load assets that have not come into view yet', (tester) async {
         await GoldenToolkit.runWithConfiguration(
           () async {
-            await tester.pumpWidgetBuilder(const ListOfImages(height: 100), surfaceSize: const Size(300, 100));
+            await tester.pumpWidgetBuilder(
+              const ListOfItemsWithOneImage(
+                itemSize: Size(100, 100),
+                cacheExtent: 1000,
+                indexThatContainsImage: 10,
+              ),
+              surfaceSize: const Size(300, 100),
+            );
             await tester.waitForAssets();
-            await tester.drag(find.byType(Scrollable), const Offset(0, -20000));
+            await tester.drag(find.byType(Scrollable), const Offset(0, -1000));
             await tester.pump();
             await expectLater(
-                find.byType(ListOfImages).first, matchesGoldenFile('goldens/list_of_images_will_show.png'));
+                find.byType(ListOfItemsWithOneImage).first, matchesGoldenFile('goldens/list_of_images_will_show.png'));
           },
           config: defaultConfiguration,
         );