+++ packages/golden_toolkit/test/image_loading/image_loading_utils.dart
@@ -35,35 +35,37 @@ GoldenToolkitConfiguration get defaultConfiguration =>
     GoldenToolkit.configuration.copyWith(primeAssets: defaultPrimeAssets);
 
 @immutable
-class ListOfImages extends StatelessWidget {
-  const ListOfImages({@required this.height});
+class ListOfItemsWithOneImage extends StatelessWidget {
+  const ListOfItemsWithOneImage({
+    @required this.itemSize,
+    @required this.indexThatContainsImage,
+    @required this.cacheExtent,
+  });
 
-  final double height;
+  final Size itemSize;
+  final int indexThatContainsImage;
+  final double cacheExtent;
 
   @override
   Widget build(BuildContext context) {
     return Container(
       color: Colors.grey,
       child: ListView.builder(
-        primary: true,
-        addAutomaticKeepAlives: false,
-        addRepaintBoundaries: true,
-        itemCount: 20,
         itemBuilder: (context, index) => Center(
           child: Row(
             children: [
               Text(index.toString()),
               const SizedBox(width: 8),
               Container(
-                width: height,
-                height: height,
+                width: itemSize.width,
+                height: itemSize.height,
                 color: Colors.lightBlue,
-                child: (index < 10) ? null : const ImageWidget(),
+                child: (index == indexThatContainsImage) ? const ImageWidget() : null,
               ),
             ],
           ),
         ),
-        cacheExtent: 2000,
+        cacheExtent: cacheExtent,
       ),
     );
   }